Clean approve transaction conponent
